renaming and organization; uniseg is non-default
