+++ contractions/filter_test.go
@@ -1,18 +1,17 @@
 package contractions_test
 
 import (
-	"strings"
 	"testing"
 
 	"github.com/clipperhouse/jargon"
 	"github.com/clipperhouse/jargon/contractions"
 )
 
-func TestSome(t *testing.T) {
+func TestContractions(t *testing.T) {
 	given := "i'll SHE’D they're Can’t should've GOTTA Wanna"
 	expected := "i will SHE WOULD they are Can not should have GOT TO Want to"
 
-	tokens := jargon.Tokenize(strings.NewReader(given))
+	tokens := jargon.TokenizeString(given)
 	got, err := contractions.Expander.Filter(tokens).String()
 	if err != nil {
 		t.Error(err)
