+++ filter.go
@@ -1,0 +1,6 @@
+package jargon
+
+// Filter is a structure for processing a stream of tokens
+type Filter interface {
+	Filter(*Tokens) *Tokens
+}
